chore: rename `instructions` -> `interpreter_loop`
This is a more descriptive name of what this module does/contains.
